+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Features
 Fixes
 
 * MultiIndexError in postprocessing if more than 2 oemof nodes are given `#174 <https://github.com/oemof/oemof-tabular/issues/174>`_
+* Updated dependencies for datapackage and tableschema
+* Error due to "fake" foreign keys in datapackage resources
 
 
 0.0.5 Patch Release - Miraculous Mary (2024-02-23)